Authenticate my mnt object and remove stale resources
I have been away from dn42 for several years for personal reasons and I
would like to re-join the network as maintaining a dn42 AS has been a
lot of fun and a great learning experience.

This commit updates my mntner and person objects with authentication in
the form of an included PGP key, claims a new ASN, a couple new netblocks
and removes an IPv4 transfer subnet which somehow was not purged from
the registry.

I am aware that the IPv4 /25 block is a large allocation, however, it is
to be used on a network with multiple branches, several of them using
dynamic addressing. (subnets will be announced aggregated, however)